+++ tests/aws/services/cloudformation/engine/test_deletion_policy.snapshot.json
@@ -1,0 +1,77 @@
+{
+  "tests/aws/services/cloudformation/engine/test_deletion_policy.py::test_deletion_policy_with_deletion": {
+    "recorded-date": "23-01-2026, 12:03:31",
+    "recorded-content": {
+      "error": {
+        "message": "An error occurred (ParameterNotFound) when calling the GetParameter operation: "
+      }
+    }
+  },
+  "tests/aws/services/cloudformation/engine/test_deletion_policy.py::test_deletion_policy_with_retain": {
+    "recorded-date": "23-01-2026, 12:03:43",
+    "recorded-content": {
+      "per-resource-events": {
+        "MyParameter": [
+          {
+            "LogicalResourceId": "MyParameter",
+            "PhysicalResourceId": "CFN-MyParameter-EZmrlQrQjjTv",
+            "ResourceStatus": "CREATE_IN_PROGRESS",
+            "ResourceType": "AWS::SSM::Parameter",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "MyParameter",
+            "PhysicalResourceId": "CFN-MyParameter-EZmrlQrQjjTv",
+            "ResourceStatus": "CREATE_COMPLETE",
+            "ResourceType": "AWS::SSM::Parameter",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "MyParameter",
+            "PhysicalResourceId": "CFN-MyParameter-EZmrlQrQjjTv",
+            "ResourceStatus": "DELETE_SKIPPED",
+            "ResourceType": "AWS::SSM::Parameter",
+            "Timestamp": "timestamp"
+          }
+        ],
+        "Stack": [
+          {
+            "LogicalResourceId": "<stack-name>",
+            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
+            "ResourceStatus": "REVIEW_IN_PROGRESS",
+            "ResourceType": "AWS::CloudFormation::Stack",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "<stack-name>",
+            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
+            "ResourceStatus": "CREATE_IN_PROGRESS",
+            "ResourceType": "AWS::CloudFormation::Stack",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "<stack-name>",
+            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
+            "ResourceStatus": "CREATE_COMPLETE",
+            "ResourceType": "AWS::CloudFormation::Stack",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "<stack-name>",
+            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
+            "ResourceStatus": "DELETE_IN_PROGRESS",
+            "ResourceType": "AWS::CloudFormation::Stack",
+            "Timestamp": "timestamp"
+          },
+          {
+            "LogicalResourceId": "<stack-name>",
+            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
+            "ResourceStatus": "DELETE_COMPLETE",
+            "ResourceType": "AWS::CloudFormation::Stack",
+            "Timestamp": "timestamp"
+          }
+        ]
+      }
+    }
+  }
+}
