--- tests/aws/services/cloudformation/engine/test_policies.snapshot.json
@@ -1,166 +1,0 @@
-{
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy[dev]": {
-    "recorded-date": "16-12-2025, 23:22:42",
-    "recorded-content": {}
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy": {
-    "recorded-date": "16-12-2025, 23:23:12",
-    "recorded-content": {}
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy_with_deletion": {
-    "recorded-date": "23-01-2026, 11:51:09",
-    "recorded-content": {
-      "error": {
-        "message": "An error occurred (ParameterNotFound) when calling the GetParameter operation: "
-      }
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy_with_retain": {
-    "recorded-date": "23-01-2026, 11:54:35",
-    "recorded-content": {
-      "per-resource-events": {
-        "MyParameter": [
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "CFN-MyParameter-q9PCAwkUG44j",
-            "ResourceStatus": "CREATE_IN_PROGRESS",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "CFN-MyParameter-q9PCAwkUG44j",
-            "ResourceStatus": "CREATE_COMPLETE",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "CFN-MyParameter-q9PCAwkUG44j",
-            "ResourceStatus": "DELETE_SKIPPED",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          }
-        ],
-        "Stack": [
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "REVIEW_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "CREATE_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "CREATE_COMPLETE",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "DELETE_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "DELETE_COMPLETE",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          }
-        ]
-      }
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestUpdatePolicyHandling::test_update_replace_policy_retain": {
-    "recorded-date": "17-12-2025, 00:30:23",
-    "recorded-content": {
-      "per-resource-events": {
-        "MyParameter": [
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "param-1a57688c",
-            "ResourceStatus": "CREATE_IN_PROGRESS",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "param-1a57688c",
-            "ResourceStatus": "CREATE_COMPLETE",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "param-d4dde898",
-            "ResourceStatus": "UPDATE_IN_PROGRESS",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "param-d4dde898",
-            "ResourceStatus": "UPDATE_COMPLETE",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "MyParameter",
-            "PhysicalResourceId": "param-1a57688c",
-            "ResourceStatus": "DELETE_SKIPPED",
-            "ResourceType": "AWS::SSM::Parameter",
-            "Timestamp": "timestamp"
-          }
-        ],
-        "Stack": [
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "REVIEW_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "CREATE_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "CREATE_COMPLETE",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "UPDATE_IN_PROGRESS",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          },
-          {
-            "LogicalResourceId": "<stack-name>",
-            "PhysicalResourceId": "arn:<partition>:cloudformation:<region>:111111111111:stack/<stack-name>/<resource:1>",
-            "ResourceStatus": "UPDATE_COMPLETE",
-            "ResourceType": "AWS::CloudFormation::Stack",
-            "Timestamp": "timestamp"
-          }
-        ]
-      }
-    }
-  }
-}
