--- tests/aws/services/cloudformation/engine/test_policies.validation.json
@@ -1,47 +1,0 @@
-{
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy": {
-    "last_validated_date": "2025-12-16T23:23:27+00:00",
-    "durations_in_seconds": {
-      "setup": 0.77,
-      "call": 20.43,
-      "teardown": 14.63,
-      "total": 35.83
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy_with_deletion": {
-    "last_validated_date": "2026-01-23T11:51:09+00:00",
-    "durations_in_seconds": {
-      "setup": 0.73,
-      "call": 14.75,
-      "teardown": 0.08,
-      "total": 15.56
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestDeletionPolicyHandling::test_deletion_policy_with_retain": {
-    "last_validated_date": "2026-01-23T11:54:35+00:00",
-    "durations_in_seconds": {
-      "setup": 0.75,
-      "call": 12.77,
-      "teardown": 0.19,
-      "total": 13.71
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestUpdatePolicyHandling::test_update_replace_policy_deletion": {
-    "last_validated_date": "2025-12-17T00:27:53+00:00",
-    "durations_in_seconds": {
-      "setup": 0.0,
-      "call": 22.19,
-      "teardown": 4.42,
-      "total": 26.61
-    }
-  },
-  "tests/aws/services/cloudformation/engine/test_policies.py::TestUpdatePolicyHandling::test_update_replace_policy_retain": {
-    "last_validated_date": "2025-12-17T00:30:27+00:00",
-    "durations_in_seconds": {
-      "setup": 1.89,
-      "call": 19.47,
-      "teardown": 4.43,
-      "total": 25.79
-    }
-  }
-}
