+++ tests/aws/services/cloudformation/engine/test_update_policy.validation.json
@@ -1,0 +1,20 @@
+{
+  "tests/aws/services/cloudformation/engine/test_update_policy.py::test_update_replace_policy_deletion": {
+    "last_validated_date": "2026-01-23T12:02:55+00:00",
+    "durations_in_seconds": {
+      "setup": 0.73,
+      "call": 23.55,
+      "teardown": 4.43,
+      "total": 28.71
+    }
+  },
+  "tests/aws/services/cloudformation/engine/test_update_policy.py::test_update_replace_policy_retain": {
+    "last_validated_date": "2026-01-23T12:03:18+00:00",
+    "durations_in_seconds": {
+      "setup": 0.0,
+      "call": 19.21,
+      "teardown": 4.51,
+      "total": 23.72
+    }
+  }
+}
